requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
celery
redis
requests
dependency-injector